AnimatedLifeBar working
Though if players gets a heal while taking damage, the bar will go to the right amount instantaneously.
That means that if he took the same amount of heal and damage, it will stop. And sometimes it will do the same if heal and damage start working at the same time. Add some sound or effects for when healing to fix it!