Opening only the next room working + more
The current room will open the door for the next room only when called to (in this case, with space).
Room properties now have ref to RangeEnemy (thought that one is not working for now.
VictoryTrigger now reacts to player.
Camera now follows player in playmode
Thief prefab was missing from projec, so it was added again